remove unused worker and parser files
