+++ webpack.config.mjs
@@ -20,13 +20,7 @@ export const WEBPACK_CONFIG = {
     },
     module: {
         rules: [
-            {
-                test: /\.ts$/,
-                use: [
-                    { loader: 'ts-loader', options: { compilerOptions: { module: 'ES2020' } } },
-                    './utils/webpack-assert-message.js',
-                ],
-            },
+            { test: /\.ts$/, use: ['ts-loader', './utils/webpack-assert-message.js'] },
             { test: /\.css$/, use: ['style-loader', 'css-loader'] },
             { test: /\.less$/, use: ['style-loader', 'css-loader', 'less-loader'] },
             { test: /\.html\.in$/, loader: 'raw-loader' },
